seguridad

+++ ClinicaUNA/src/main/java/cr/ac/una/clinicauna/util/Request.java
@@ -8,6 +8,10 @@ package cr.ac.una.clinicauna.util;
 /*import jakarta.json.Json;
 import jakarta.json.JsonObject;
 import jakarta.json.JsonReader;*/
+import cr.ac.una.clinicauna.service.UserService;
+import jakarta.json.Json;
+import jakarta.json.JsonObject;
+import jakarta.json.JsonReader;
 import jakarta.ws.rs.client.Client;
 import jakarta.ws.rs.client.ClientBuilder;
 import jakarta.ws.rs.client.Entity;
@@ -18,6 +22,8 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.MultivaluedHashMap;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.core.Response;
+import java.io.StringReader;
+import java.util.Base64;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -64,7 +70,7 @@ public class Request {
         this.builder = webTarget.request(MediaType.APPLICATION_JSON);
         MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
         headers.add("Content-Type", "application/json; charset=UTF-8");
-       if (AppContext.getInstance().get("Token") != null) {
+        if (AppContext.getInstance().get("Token") != null) {
             headers.add("Authorization", AppContext.getInstance().get("Token").toString());
         }
         builder.headers(headers);
@@ -94,31 +100,42 @@ public class Request {
     }
 
     public void get() {
-        // TODO
-        response = builder.get();
+        if (verifyTokenExp()) {
+            response = builder.get();
+        }
     }
 
     public void getToken() {
         response = builder.get();
     }
 
-    //TODO
+    public void getRenewal() {
+        JsonObject payload = getPayloadToken(AppContext.getInstance().get("Token").toString());
+        MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
+        headers.add("Authorization", payload.getString("rnt"));
+        setHeader(headers);
+        response = builder.get();
+    }
 
+    //TODO
     public void post(Object clazz) {
-        //TODO
+        if (verifyTokenExp()) {
             Entity<?> entity = Entity.entity(clazz, "application/json; charset=UTF-8");
             response = builder.post(entity);
+        }
     }
 
     public void put(Object clazz) {
-        // TODO
+        if (verifyTokenExp()) {
             Entity<?> entity = Entity.entity(clazz, "application/json; charset=UTF-8");
             response = builder.put(entity);
+        }
     }
 
     public void delete() {
-        // TODO
+        if (verifyTokenExp()) {
             response = builder.delete();
+        }
     }
 
     public int getStatus() {
@@ -126,9 +143,9 @@ public class Request {
     }
 
     public Boolean isError() {
-        if(getStatus()== Response.Status.UNAUTHORIZED.getStatusCode()){
-            new Thread(){
-                public void run(){
+        if (getStatus() == Response.Status.UNAUTHORIZED.getStatusCode()) {
+            new Thread() {
+                public void run() {
                     try {
                         Thread.sleep(4000);
                         Platform.runLater(new Runnable() {
@@ -179,7 +196,44 @@ public class Request {
     }
 
     // TODO
-
+    private boolean verifyTokenExp() {
+        if (AppContext.getInstance().get("Token") == null) {
+            return true;//No se ha logeado.
+        }
+        JsonObject payload = getPayloadToken(AppContext.getInstance().get("Token").toString());
+        if (payload.getJsonNumber("exp").longValue() > System.currentTimeMillis() / 1000) {
+            return true;
+        } else {
+            payload = getPayloadToken(payload.getString("rnt"));
+            if (payload != null && payload.getJsonNumber("exp").longValue() > System.currentTimeMillis() / 1000) {
+                UserService empleadoService = new UserService();
+                Respuesta respuesta = empleadoService.renovarToken();
+                if (respuesta.getEstado()) {
+                    AppContext.getInstance().set("Token", respuesta.getResultado("Token").toString());
+                    MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
+                    headers.add("Authorization", respuesta.getResultado("Token").toString());
+                    setHeader(headers);
+                    return true;
+                }
+            }
+            response = Response.status(401, "El token a expirado.").build();
+            return false;
+        }
+    }
     // TODO
 
+    private JsonObject getPayloadToken(String token) {
+        if (token != null && !token.isEmpty()) {
+            token = token.substring(AUTHENTICATION_SCHEME.length()).trim();
+            String[] parts = token.split("\\.");
+            String decodedToken = new String(Base64.getUrlDecoder().decode(parts[1]));
+            JsonObject object;
+            try (JsonReader jsonReader = Json.createReader(new StringReader(decodedToken))) {
+                object = jsonReader.readObject();
+            }
+            return object;
+        } else {
+            return null;
+        }
+    }
 }
